requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
pygame>=2.5.0
pyyaml>=6.0
